test: extract issue-export suite from AppStateTests (slice 1 of #431)
Moves the 10 issue export & review tests (canExportIssues,
exportSelectedIssues, confirmPendingExportReview) verbatim out of the
2,347-line AppStateTests into a focused AppStateIssueExportTests suite.
No assertions changed; total test count unchanged (91 -> 81 + 10). These
tests depend only on the shared AppStateHarness, so no helpers moved.

Remaining clusters (session-library, permissions, retry, tracker-config)
stay in AppStateTests for follow-up slices tracked on #431.

Validation: build OK, both suites green (91 tests, 0 failures), guardrails pass.

Refs #431
